restructure editor, export serializer instead of plugins
